@Sean-Der 2019-06-03
We need to evaluate performance across all of `pion` and go from there, closing for now since it conflicts.